+++ test/TestAsset/FooCommand.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * @license   http://opensource.org/licenses/BSD-3-Clause BSD-3-Clause
+ * @copyright Copyright (c) 2016 Zend Technologies USA Inc. (http://www.zend.com)
+ */
+
+namespace ZFTest\Console\TestAsset;
+
+class FooCommand
+{
+    public function __invoke()
+    {
+    }
+}